Provenance rows wear recency colours; rail icon sizing; network-panel consistency
Three polish items (the user 2026-07-27):

- The age-provenance popover colours every timed line — timestamp AND
  text — by its own recency, the chat tab-hover treatment, off the shared
  colormap. The ramp that fleet.ts had copied verbatim from render.ts is
  extracted to ui/webview/age-color.ts (the popover would have been a
  third copy); fleet imports it, render.ts still carries its
  settings-entangled twin with a keep-in-sync pointer. ProvRow gains the
  row's own epoch so the renderer can tint without re-parsing the label.

- Rail cluster: the bell draws a touch smaller (17px, lighter stroke) and
  the ↻ text glyph spans the full 18px icon-row height like its svg
  neighbors (line-height pinned so it can't grow the bar), on both the
  desktop rail and the mobile bar.

- The network panel's buttons join the settings modal's action-button
  vocabulary: row buttons at 12px (they inherited the panel's 13px body
  size — the one size that matched nothing) with that family's hover;
  the "Previously attached" header takes .rs-sec's exact treatment
  (700 / .08em); the empty-state gray becomes the panel's one dim gray.
